LayoutTests/fast/repaint/drop-shadow-writing-modes-expected.txt

Test repaint rects when the drop-shadow filter property changes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS layerTreeAsText.indexOf('(rect 60 160 156 156)') > -1 is true
PASS successfullyParsed is true

TEST COMPLETE
here is some text
